ENH: en el metode que retorna l'espai lliure del disc dur s'ha canviat la struc statfs, que estava deprecated per statvfs, tambe s'han canviat els tipus utilitzats per unsigned long long

+++ starviewer/src/inputoutput/cachepool.h
@@ -20,7 +20,7 @@ class CachePool{
 public:
 
     CachePool();
-    double getFreeTotalSpace();
+    unsigned long long getFreeTotalSpace();
 
     void removeStudy(std::string studyUID);
